Create reusable modal component
