--- packages/cpt-ui/assets/styles/EpsLogoutModal.scss
@@ -1,43 +1,0 @@
-.EpsLogoutModalOverlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    height: 100vh;
-    width: 100vw;
-    background-color: rgba(0, 0, 0, 0.5); // darken the background
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    z-index: 9999;
-  }
-  
-  .EpsLogoutModalContent {
-    background: #fff;
-    padding: 2rem;
-    border-radius: 4px;
-    position: relative;
-    max-width: 600px;
-    width: 90%;
-  }
-  
-  .EpsLogoutModalButton {
-    width: 40%;
-  }
-
-  .EpsLogoutModalCloseButton {
-    position: absolute;
-    top: 1rem;
-    right: 1rem;
-    font-size: 1.5rem;
-    background: transparent;
-    border-radius: 4px;
-    cursor: pointer; // TODO: Does this want to be a normal cursor Or a pointer?
-  }
-  
-  .EpsLogoutModalButtonGroup {
-    display: flex;
-    justify-content: center;
-    gap: 20%;
-    margin-top: 2rem;
-  }
-
