+++ packages/cpt-ui/assets/styles/EpsModal.scss
@@ -1,0 +1,42 @@
+.eps-modal-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  height: 100vh;
+  width: 100vw;
+  background-color: rgba(0, 0, 0, 0.5); // darken the background
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 9999;
+}
+
+.eps-modal-content {
+  background: #fff;
+  padding: 2rem;
+  border-radius: 4px;
+  position: relative;
+  max-width: 700px;
+  width: 90%;
+}
+
+.eps-modal-button {
+  width: 40%;
+}
+
+.eps-modal-close-button {
+  position: absolute;
+  top: 1rem;
+  right: 1rem;
+  font-size: 1.5rem;
+  background: transparent;
+  border-radius: 4px;
+  cursor: pointer; // TODO: Does this want to be a normal cursor Or a pointer?
+}
+
+.eps-modal-button-group {
+  display: flex;
+  justify-content: center;
+  gap: 20%;
+  margin-top: 2rem;
+}
