+++ packages/cpt-ui/components/EpsLogoutModal.tsx
@@ -1,8 +1,8 @@
 "use client";
-import React, { useEffect } from "react";
+import React from "react";
 import { Button } from "nhsuk-react-components"
 
-import "@/assets/styles/EpsLogoutModal.scss";
+import { EpsModal } from "@/components/EpsModal";
 import { EpsLogoutModalStrings } from "@/constants/ui-strings/EpsLogoutModalStrings";
 
 interface EpsLogoutModalProps {
@@ -12,68 +12,28 @@ interface EpsLogoutModalProps {
 }
 
 export function EpsLogoutModal({ isOpen, onClose, onConfirm }: EpsLogoutModalProps) {
-  // Close modal on `Escape` key
-  useEffect(() => {
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if (e.key === "Escape") {
-        onClose();
-      }
-    };
-
-    window.addEventListener("keydown", handleKeyDown);
-    return () => window.removeEventListener("keydown", handleKeyDown);
-  }, [onClose]);
-  
-  // Close if user clicks outside the modal content
-  const handleBackdropClick = (e: React.MouseEvent) => {
-    if (e.target === e.currentTarget) {
-      onClose();
-    }
-  };
-  
-  // Close if user activates on the background
-  const handleBackdropActivate = (e: React.KeyboardEvent) => {
-    if (e.key === "Enter" || e.key === " ") {
-      onClose();
-    }
-  };
-
-  // If the modal isn’t open, don’t render anything
-  if (!isOpen) return null;
   
   return (
-    <div
-      className="EpsLogoutModalOverlay" 
-      onClick={handleBackdropClick}
-      role="button"
-      tabIndex={0}
-      onKeyDown={handleBackdropActivate}
-    >
-      <div className="EpsLogoutModalContent" role="dialog" aria-modal="true">
-        <button className="EpsLogoutModalCloseButton" onClick={onClose} aria-label="Close modal">
-          ×
-        </button>
-        
+    <EpsModal isOpen={isOpen} onClose={onClose}>
         <h2>{EpsLogoutModalStrings.title}</h2>
         <p>{EpsLogoutModalStrings.caption}</p>
         
         {/* TODO: style appropriately */}
-        <div className="EpsLogoutModalButtonGroup">
+        <div className="eps-modal-button-group">
           <Button 
-            className="nhsuk-button EpsLogoutModalButton" 
+            className="nhsuk-button eps-modal-button" 
             onClick={onConfirm}
           >
             {EpsLogoutModalStrings.confirmButtonText}
           </Button>
           
           <Button 
-            className="nhsuk-button nhsuk-button--secondary EpsLogoutModalButton" 
+            className="nhsuk-button nhsuk-button--secondary eps-modal-button" 
             onClick={onClose}
           >
             {EpsLogoutModalStrings.cancelButtonText}
           </Button>
         </div>
-      </div>
-    </div>
+    </EpsModal>
   );
 }
