+++ packages/cpt-ui/constants/ui-strings/EpsLogoutModalStrings.ts
@@ -1,6 +1,6 @@
 export const EpsLogoutModalStrings = {
   title: "Are you sure you want to log out?",
   caption: "Logging out will end your session.",
-  confirmButtonText: "Yes",
-  cancelButtonText: "No"
+  confirmButtonText: "Log out",
+  cancelButtonText: "Cancel"
 }
